requirements.txt
Django==4.1.4
djangorestframework==3.14.0
djoser==2.2.0
djangorestframework_simplejwt==5.3.0
psycopg2_binary==2.9.8
docker==24.0.6
python_dotenv==1.0.0